API: Rework mechanism of dynamically enabling/disabling components.
Previously one could pass a condition which is backed by a Boolean provider to determine the enabled/visibility state of the row/cell. This wouldn't work though as even though the Condition interface promised to support listening for changes the implementations did not meet this standard.

We introduce a separate BoundCondition which explicitly offers this functionality and differentiate these two cases. The #enableIf and #visibleIf functions are only allowed to take in BoundConditions.

To make it easier to provide such bound Conditions one can create them from an Observable instance. We also provide convenience methods to get the state of a ui component.