API: Add support for dynamic text as ObservableProperty<String>.
If you want to use TextLabel etc in a context where they are used as a renderer contained in a visual hierarchy (i.e. not as a renderer for tables, trees etc.) the label needs to be aware its content may change any time without the Text actually sending a notification. PseudoObservablyProperty simply is a wrapper around a provider and code needs to be aware it uses a pseudo observable if it supports them.